+++ NEWS.md
@@ -7,6 +7,7 @@
 * Added `ShowHCAProjects` to show all available Human Cell Atlas projects.
 * Added `ShowCELLxGENEDatasets` to show all available CELLxGENE datasets.
 * Added `StatDBAttribute`to stat database attributes.
+* Added `ParseHCA` to download objects from Human Cell Atlas.
 
 -------------------